Send websocket messages immediately in all cases
This is important as we rely on websockets to show progress and
updates, and scheduling a greenlet for sending a message, messes with that as we
don't know when that greenlet will be scheduled to run.
