+++ rotkehlchen/user_messages.py
@@ -73,6 +73,13 @@ class MessagesAggregator():
             message_type: WSMessageType,
             data: Union[dict[str, Any], list[Any]],
     ) -> None:
+        """Sends a websocket message
+
+        Specify its type and data.
+
+        `wait_on_send` is used to determine if the message should be sent asynchronously
+        by spawning a greenlet or if it should just do it synchronously.
+        """
         fallback_msg = json.dumps({'type': str(message_type), 'data': data})  # noqa: E501  # kind of silly to repeat it here. Same code in broadcast
 
         if self.rotki_notifier is not None:
